BUGGED: seperated traitement file in action and semaine
